refactor: Rating styles and functions to get these should be available globally
